게시판만들기

+++ board/src/main/java/org/iclass/board/mapper/CommunityCommentsMapper.java
@@ -7,7 +7,7 @@ import org.iclass.board.dto.CommunityCommentsDto;
 
 @Mapper
 public interface CommunityCommentsMapper {
-    CommunityCommentsDto selectByPk(Integer idx);
+    CommunityCommentsDto selectByPk(long idx);
 
     List<CommunityCommentsDto> selectAll();
